# docs/endpoints/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ithina Admin Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/platform-users": {
      "get": {
        "tags": [
          "platform-users"
        ],
        "summary": "List platform users",
        "description": "List Ithina staff users. Gated on ADMIN.USERS.VIEW.GLOBAL — TENANT JWTs and PLATFORM users lacking the grant receive 403 PERMISSION_DENIED. Supports filter by status, case-insensitive search across email/full_name, sort, and offset/limit pagination. Each item carries an inline `roles[]` array (Step 6.8.3 augmentation); platform users' assignments have no org-node anchor so `org_node_id` / `org_node_name` are always null.",
        "operationId": "list_platform_users_api_v1_platform_users_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PlatformUserStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status: INVITED, ACTIVE, or SUSPENDED.",
              "title": "Status"
            },
            "description": "Filter by status: INVITED, ACTIVE, or SUSPENDED."
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive substring match across email and full_name.",
              "title": "Search"
            },
            "description": "Case-insensitive substring match across email and full_name."
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort key: created_at_asc, created_at_desc, full_name_asc, full_name_desc, email_asc, email_desc.",
              "default": "created_at_desc",
              "title": "Sort"
            },
            "description": "Sort key: created_at_asc, created_at_desc, full_name_asc, full_name_desc, email_asc, email_desc."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Pagination offset.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Pagination offset."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Pagination limit (max 200).",
              "default": 50,
              "title": "Limit"
            },
            "description": "Pagination limit (max 200)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformUserListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/platform-users/{user_id}": {
      "get": {
        "tags": [
          "platform-users"
        ],
        "summary": "Get platform user by ID",
        "description": "Get a single platform user by UUID. Gated on ADMIN.USERS.VIEW.GLOBAL — TENANT JWTs and PLATFORM users lacking the grant receive 403 PERMISSION_DENIED. Unknown user_id returns 404 PLATFORM_USER_NOT_FOUND. Includes inline `roles[]` array (Step 6.8.3 augmentation).",
        "operationId": "get_platform_user_api_v1_platform_users__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformUserRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Pagination": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "total",
          "offset",
          "limit"
        ],
        "title": "Pagination",
        "description": "Pagination metadata block. Carried only by list responses."
      },
...
      "PlatformUserListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlatformUserRead"
            },
            "type": "array",
            "title": "Items"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "type": "object",
        "required": [
          "items",
          "pagination"
        ],
        "title": "PlatformUserListResponse",
        "description": "List endpoint response envelope: {items, pagination} per D-30."
      },
      "PlatformUserRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "status": {
            "$ref": "#/components/schemas/PlatformUserStatus"
          },
          "invited_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invited At"
          },
          "invitation_accepted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invitation Accepted At"
          },
          "suspended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suspended At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/UserRoleAssignmentItem"
            },
            "type": "array",
            "title": "Roles"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "full_name",
          "status",
          "invited_at",
          "invitation_accepted_at",
          "suspended_at",
          "created_at",
          "updated_at",
          "roles"
        ],
        "title": "PlatformUserRead",
        "description": "Platform (Ithina staff) user as returned by the API.\n\nAudit-actor IDs and ``auth0_sub`` are intentionally absent: those\nare internal-only. The frontend renders lifecycle state from the\ntimestamp fields alone.\n\n``roles`` (Step 6.8.3): inline list of role assignments from\n``platform_user_role_assignments``. Each item's ``org_node_id``\nand ``org_node_name`` are always null (platform-side assignments\nhave no org-node anchor); the keys are still present so the wire\nshape stays uniform with tenant-side. Always present; empty array\n(not null) for platform users with no assignments. Both ACTIVE\nand INACTIVE assignments included."
      },
      "PlatformUserStatus": {
        "type": "string",
        "enum": [
          "INVITED",
          "ACTIVE",
          "SUSPENDED"
        ],
        "title": "PlatformUserStatus",
        "description": "Lifecycle states for a platform user. Mirrors ``platform_user_status_enum``."
      },
...
      "UserRoleAssignmentItem": {
        "properties": {
          "assignment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Assignment Id"
          },
          "role_id": {
            "type": "string",
            "format": "uuid",
            "title": "Role Id"
          },
          "role_name": {
            "type": "string",
            "title": "Role Name"
          },
          "role_code": {
            "type": "string",
            "title": "Role Code"
          },
          "status": {
            "$ref": "#/components/schemas/UserRoleAssignmentStatus"
          },
          "granted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Granted At"
          },
          "org_node_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Org Node Id"
          },
          "org_node_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Org Node Name"
          }
        },
        "type": "object",
        "required": [
          "assignment_id",
          "role_id",
          "role_name",
          "role_code",
          "status",
          "granted_at",
          "org_node_id",
          "org_node_name"
        ],
        "title": "UserRoleAssignmentItem",
        "description": "Inline role-assignment item for user response augmentation.\n\nUsed by both tenant_users and platform_users responses. For\nplatform users, ``org_node_id`` and ``org_node_name`` are always\nNone (the underlying ``platform_user_role_assignments`` table has\nno org-node anchoring). The two keys are present in the JSON\noutput regardless of audience so the wire shape is uniform.\n\nAll assignments are returned regardless of status (ACTIVE +\nINACTIVE both ship); frontend filters as needed.\n\nSchema home: ``schemas/tenant_user.py`` (re-exported from\n``schemas/platform_user.py``). Distinct from\n``schemas/role_assignment.py`` shapes which serve the standalone\n``/role-assignments`` endpoint with a richer nested envelope."
      },
      "UserRoleAssignmentStatus": {
        "type": "string",
        "enum": [
          "ACTIVE",
          "INACTIVE"
        ],
        "title": "UserRoleAssignmentStatus",
        "description": "Mirrors ``user_role_assignment_status_enum``.\n\nShared by both physical assignment tables. ACTIVE rows count\ntoward role usage (E1's ``user_count`` aggregate); INACTIVE rows\nrepresent revoked assignments retained for audit history."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}